refactor(web): extract shared task tool helpers
